Fix multiple bugs and problems
I know, this should be more commits but each time I tried to fix one
bug, another came up and I wasn't sure if a fix for the one problem does
not raise up new problems. So I had to fix them all in one and now it
would be a lot of work to split this commit into multiple ones.

Sorry for that.

Now, here is the list of changes/fixes:

- trailing slashes of directory names are removed
  In some environments (for me: ssh in chroot) the home directory
  stored in $HOME had a trailing slash, in other environments (for me:
  my normal desktop) there was no trailing slash. This showed very
  different behaviour between the two environments.
- the config applyment just skips the home directory but continues to
  search in the home's parent directories for configs to apply.
- another parameter `rootpath` is added in order to tell the script a
  root path where it should stop looking for configs. This was needed to
  successfully run the tests. Otherwise the plugin also sourced my
  config in my home directory because I was running the tests in a
  subdirectory of my home dir.